requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
dateparser==0.6.0
future==0.16.0
futures==3.2.0
html5lib==0.999999999
lxml==4.1.1
mysql-connector==2.1.6
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyMySQL==0.7.11
pypiwin32==219
python-dateutil==2.6.1
pytz==2018.4
regex==2017.9.23
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
sqlacodegen==1.1.6
urllib3==1.22
